Added session timeout
Edited session_suth(cookies) function to check if current session has expired. Changed the expiry time to 30 minutes from when session begins (to test we can change this to minutes=0.5 so it timesout after 30 seconds- I have tested it and it worked ok for me) 
I also edited the response for /logout and def profile() to redirect to the login page rather than rendering the login.html page so that the csrf token gets created via these two routes.